Make it possible to adjust the name of the resource for pullResource and pushResource.
Previously the plugin would use the filename with extension and with all '.' characters
removed as a resource name, now one can optionally provide a 'resource_name' function
via the global options argument. This function is invoked with the Vinyl object as single
parameter.
Similarly, a 'target_name' function can be provided to adjust the default output path
for files produced with 'pullResource'.