Allow to specify translation mode query param when pulling translations
This can be useful when dealing with empty translations for files that are invalid when translation is listed but empty. For example, XLIFF <target> element should always contain something or the translation shouldn't be listed at all. Transifex API always seems to send the translation even when it's empty (with <target> missing completely) and as a workaround ?mode=sourceastranslation can be used to avoid blank translations. If specified, <target> will be present holding original translation.